feat: 增添全链路追踪功能 traceId (#3)
* feat: 基于 BeanPostProcessor 扩展思想轻松应对后续变化


* chore: 增添 hutool 工具包，更新 README 文档

* feat: 增添 hutool 工具包，使用雪花算法生成业务ID

* feat: 对 PostProcessor 做了些修正。调整注释与代码实现

* feat: 添加全链路 traceId、美化输入日志。给出各模块的联合协作简易 demo

---------

+++ polaris-app/src/main/java/com/achobeta/config/ThreadPoolConfigProperties.java
@@ -12,7 +12,7 @@ public class ThreadPoolConfigProperties {
     /** 最大线程数 */
     private Integer maxPoolSize = 200;
     /** 最大等待时间 */
-    private Long keepAliveTime = 10L;
+    private Integer keepAliveTime = 10;
     /** 最大队列数 */
     private Integer blockQueueSize = 5000;
     /*
